requirements.txt
telethon>=1.34.0
aiohttp
python-dotenv
